refactor: split abstractions and implementations (#247)
Why?
When support for platforms other than Rekordbox is added,
the pre-existing modules for Collections, Playlists, and Tracks
will quickly become too verbose.
