+++ docs/reference/collection/index.md
@@ -1,11 +1,14 @@
 # Collection
 
 ::: djtools.collection.config
-::: djtools.collection.collections
-::: djtools.collection.playlists
-::: djtools.collection.tracks
 ::: djtools.collection.playlist_builder
 ::: djtools.collection.playlist_filters
+::: djtools.collection.base_collection
+::: djtools.collection.rekordbox_collection
+::: djtools.collection.base_playlist
+::: djtools.collection.rekordbox_playlist
+::: djtools.collection.base_track
+::: djtools.collection.rekordbox_track
 ::: djtools.collection.shuffle_playlists
 ::: djtools.collection.copy_playlists
 ::: djtools.collection.helpers
