+++ src/djtools/collection/__init__.py
@@ -1,5 +1,7 @@
 """The `collection` package contains modules:
-    * `collections`: abstractions and implementations for collections
+    * `base_collection`: abstraction for Collection
+    * `base_playlist`: abstraction for Playlist
+    * `base_track`: abstraction for Track
     * `config`: the configuration object for the `collection` package
     * `copy_playlists`: copies audio files for tracks within a set of
         playlists to a new location and writes a new collection with these
@@ -11,17 +13,20 @@
         `collection_playlists.yaml`
     * `playlist_filters`: abstractions and implementations for playlist filters
     * `playlists`: abstractions and implementations for playlists
+    * `rekordbox_collection`: implementation of Collection for Rekordbox
+    * `rekordbox_playlist`: implementation of Playlist for Rekordbox
+    * `rekordbox_track`: implementation of Track for Rekordbox
     * `shuffle_playlists`: writes sequential numbers to tags of shuffled tracks
         in playlists to emulate playlist shuffling
     * `tracks`: abstractions and implementations for tracks
 """
 
-from djtools.collection.collections import RekordboxCollection
 from djtools.collection.copy_playlists import copy_playlists
 from djtools.collection.playlist_builder import collection_playlists
-from djtools.collection.playlists import RekordboxPlaylist
+from djtools.collection.rekordbox_collection import RekordboxCollection
+from djtools.collection.rekordbox_playlist import RekordboxPlaylist
+from djtools.collection.rekordbox_track import RekordboxTrack
 from djtools.collection.shuffle_playlists import shuffle_playlists
-from djtools.collection.tracks import RekordboxTrack
 
 
 COLLECTION_OPERATIONS = {
