+++ tests/collection/test_base_collection.py
@@ -1,0 +1,15 @@
+"""Testing for the collection module."""
+import pytest
+
+from djtools.collection.base_collection import Collection
+
+
+def test_collection_raises_type_error():
+    """Test Collection class."""
+    with pytest.raises(
+        TypeError,
+        match=(
+            "Can't instantiate abstract class Collection with abstract method"
+        ),
+    ):
+        Collection(path="")
