+++ tests/collection/test_rekordbox_playlist.py
@@ -1,18 +1,7 @@
 """Testing for the playlists module."""
 import pytest
 
-from djtools.collection.playlists import Playlist, RekordboxPlaylist
-
-
-def test_playlist_raises_type_error():
-    """Test Playlist class."""
-    with pytest.raises(
-        TypeError,
-        match=(
-            "Can't instantiate abstract class Playlist with abstract method"
-        ),
-    ):
-        Playlist()
+from djtools.collection.rekordbox_playlist import RekordboxPlaylist
 
 
 def test_rekordboxplaylist_getitem(rekordbox_playlist):
